firecracker: Set --enable-pci option when available
This enables the PCI virtio transport which is supposed to be faster than the MMIO one.

Support for this flag is currently only availble in Nixpkgs unstable.

I haven't been able to measure these perf improvements.

However this is still what the Firecracker docs say to do.

An option for this can be added as needed, for now just keep it simple.